requirements.txt
Django==1.9.10
django-appconf==1.0.2
django-bootstrap-ui==0.2.0
django-compressor==2.1
django-tag-parser==2.1
dominate==2.1.17
psycopg2==2.6.2
rcssmin==1.0.6
rjsmin==1.0.12
uWSGI==2.0.14

# fab is used for deployment
# dependent: asn1crypto-0.22.0 bcrypt-3.1.3 cffi-1.11.0 cryptography-2.0.3 idna-2.6 paramiko-2.3.0
# pyasn1-0.3.6 pycparser-2.18 pynacl-1.1.2 six-1.11.0
Fabric3==1.13.1.post1
